Improved Slide Content Capture
Added logic to only capture the slide content and correctly set the ratio & size of the PDF. This improves and resolves 2 issues:

a. If the window was not the same aspect ratio as the saved PDF (1920 x 980) the PDF content was squished or stretched to fit the hard coded aspect ratio.

b. The captureVisibleTab captured the entire window including all the control/navigation chrome and any black borders at top or bottom (from the window not 100% matching the slide aspect ratio).